--- cmd/commandline/test.go
@@ -1,45 +1,0 @@
-package main
-
-import "github.com/spf13/cobra"
-
-/*
- Test is a very important component of Dify plugins, to ensure every plugin is working as expected
- We must provide a way to make the test a pipeline and use standard CI/CD tools to run the tests
-
- However, it's hard to make it friendly to tests author to write the codes,
-*/
-
-var (
-	testZipPluginValidateCommand = &cobra.Command{
-		Use:   "validate-pkg",
-		Short: "Validate a zip plugin",
-		Long:  "Validate a zip plugin",
-		Args:  cobra.ExactArgs(1),
-		Run: func(c *cobra.Command, args []string) {
-			// validate if the pkg is valid
-		},
-	}
-
-	testInvokeToolPluginCommand = &cobra.Command{
-		Use:   "invoke-tool",
-		Short: "Invoke a tool",
-		Long:  "Invoke a tool",
-		Args:  cobra.ExactArgs(1),
-		Run: func(c *cobra.Command, args []string) {
-			// invoke a tool
-		},
-	}
-
-	testValidateToolProviderCommand = &cobra.Command{
-		Use:   "validate-tool-provider",
-		Short: "Validate a tool provider",
-		Long:  "Validate a tool provider",
-		Args:  cobra.ExactArgs(1),
-		Run: func(c *cobra.Command, args []string) {
-			// validate a tool provider
-		},
-	}
-)
-
-func init() {
-}
